move sample data to sample_data.py
